Move petals.dht_utils -> petals.utils.dht
