+++ src/petals/utils/__init__.py
@@ -4,3 +4,4 @@ from petals.utils.auto_config import (
     AutoDistributedModelForCausalLM,
     AutoDistributedModelForSequenceClassification,
 )
+from petals.utils.dht import declare_active_modules, get_remote_module_infos
